+++ qlib/contrib/backtest/env.py
@@ -5,13 +5,13 @@ import json
 import copy
 import warnings
 import pathlib
+import numpy as np
 import pandas as pd
-from loguru import Logger
-from ...data import D, Cal
-from ...utils import get_date_in_file_name
-from ...utils import get_pre_trading_date
-from ..backtest.order import Order
-from ..utils import init_instance_by_config
+from ...data.data import Cal
+from ...utils import get_sample_freq_calendar
+from .order import Order
+
+
 class TradeCalendarBase:
 
     def _reset_trade_calendar(self, start_time, end_time):
@@ -20,10 +20,10 @@ class TradeCalendarBase:
         if end_time:
             self.end_time = pd.Timestamp(end_time)
         if self.start_time and self.end_time:
-            _calendar, freq, freq_sam = get_sample_freq_calendar(freq=step_bar)
+            _calendar, freq, freq_sam = get_sample_freq_calendar(freq=self.step_bar)
             self.calendar = _calendar
             _start_time, _end_time, _start_index, _end_index = Cal.locate_index(self.start_time, self.end_time, freq=freq, freq_sam=freq_sam)
-            _trade_calendar = self.calendar[_start_index, _end_index + 1]
+            _trade_calendar = self.calendar[_start_index: _end_index + 1]
             if _start_time != self.start_time:
                 self.trade_calendar = np.hstack((self.start_time, _trade_calendar, self.end_time))
                 self.start_index = _start_index - 1
@@ -40,7 +40,7 @@ class TradeCalendarBase:
         trade_index = trade_index - shift
         if 0 < trade_index < self.trade_len - 1: 
             trade_start_time = self.trade_calendar[trade_index - 1]
-            trade_end_time = self.trade_calendar[trade_index] - pd.Timestamp(second=1)
+            trade_end_time = self.trade_calendar[trade_index] - pd.Timedelta(seconds=1)
             return trade_start_time, trade_end_time
         elif trade_index == self.trade_len - 1:
             trade_start_time = self.trade_calendar[trade_index - 1]
@@ -68,7 +68,7 @@ class BaseEnv(TradeCalendarBase):
         end_time=None,
         trade_account=None,
         verbose=False,
-        **kwargs
+        **kwargs,
     ):
         self.step_bar = step_bar
         self.verbose = verbose
@@ -76,24 +76,24 @@ class BaseEnv(TradeCalendarBase):
 
     def _get_position(self):
         return self.trade_account.current
-
     
 
     def reset(self, start_time=None, end_time=None, trade_account=None, **kwargs):
         if start_time or end_time:
             self._reset_trade_calendar(start_time=start_time, end_time=end_time)
-        self.trade_account = trade_account
+        if trade_account:
+            self.trade_account = trade_account
         
         for k, v in kwargs:
             if hasattr(self, k):
                 setattr(self, k, v)
 
-    def get_first_state(self):
+    def get_init_state(self):
         init_state = {"current": self._get_position()}
         return init_state
     
 
-    def execute(self, order_list, **kwargs):
+    def execute(self, order_list=None, **kwargs):
         self.trade_index = self.trade_index + 1
 
     def finished(self):
@@ -122,13 +122,13 @@ class SplitEnv(BaseEnv):
         #if self.track:
         #    yield action
         #episode_reward = 0
-        super(SimulatorEnv, self).execute(**kwargs)
+        super(SplitEnv, self).execute(**kwargs)
         trade_start_time, trade_end_time = self._get_trade_time(trade_index=self.trade_index)
         self.sub_env.reset(start_time=trade_start_time, end_time=trade_end_time, trade_account=self.trade_account)
         self.sub_strategy.reset(start_time=trade_start_time, end_time=trade_end_time, trade_order_list=order_list)
         trade_state = self.sub_env.get_init_state()
         while not self.sub_env.finished():
-            _order_list = self.sub_strategy.generate_order(**trade_state)
+            _order_list = self.sub_strategy.generate_order_list(**trade_state)
             trade_state, trade_info = self.sub_env.execute(order_list=_order_list)
             #episode_reward += sub_reward
         _obs = {"current": self._get_position()}
@@ -149,11 +149,12 @@ class SimulatorEnv(BaseEnv):
         verbose=False,
         **kwargs,
     ):
-        super(SimulatorEnv, self).__init__(step_bar=step_bar, start_time=start_time, end_time=end_time, trade_account=trade_account, trade_exchange=trade_exchange, verbose=verbose)
+        super(SimulatorEnv, self).__init__(step_bar=step_bar, start_time=start_time, end_time=end_time, trade_account=trade_account, trade_exchange=trade_exchange, verbose=verbose, **kwargs)
 
-    def reset(trade_exchange=None, **kwargs):
+    def reset(self, trade_exchange=None, **kwargs):
         super(SimulatorEnv, self).reset(**kwargs)
-        self.trade_exchange=trade_exchange
+        if trade_exchange:
+            self.trade_exchange=trade_exchange
 
     def execute(self, order_list, **kwargs):
         """
@@ -162,7 +163,7 @@ class SimulatorEnv(BaseEnv):
         if self.finished():
             raise StopIteration(f"this env has completed its task, please reset it if you want to call it!")
         super(SimulatorEnv, self).execute(**kwargs)
-        ttrade_start_time, trade_end_time = self._get_trade_time(trade_index=self.trade_index)
+        trade_start_time, trade_end_time = self._get_trade_time(trade_index=self.trade_index)
         trade_info = []
         for order in order_list:
             if self.trade_exchange.check_order(order) is True:
