+++ qlib/contrib/backtest/interpreter.py
@@ -1,0 +1,15 @@
+
+class BaseInterpreter:
+    @staticmethod
+    def interpret(**kwargs):
+        raise NotImplementedError("interpret is not implemented!")
+
+class ActionInterpreter:
+    @staticmethod
+    def interpret(action, **kwargs):
+        return action
+
+class StateInterpreter:
+    @staticmethod
+    def interpret(state, **kwargs):
+        return state
