+++ qlib/contrib/backtest/report.py
@@ -21,20 +21,20 @@ class Report:
         self.costs = OrderedDict()  # trade cost for each trade date
         self.values = OrderedDict()  # value for each trade date
         self.cashes = OrderedDict()
-        self.latest_report_date = None  # pd.TimeStamp
+        self.latest_report_time = None  # pd.TimeStamp
 
     def is_empty(self):
         return len(self.accounts) == 0
 
     def get_latest_date(self):
-        return self.latest_report_date
+        return self.latest_report_time
 
     def get_latest_account_value(self):
-        return self.accounts[self.latest_report_date]
+        return self.accounts[self.latest_report_time]
 
     def update_report_record(
         self,
-        trade_date=None,
+        trade_time=None,
         account_value=None,
         cash=None,
         return_rate=None,
@@ -44,7 +44,7 @@ class Report:
     ):
         # check data
         if None in [
-            trade_date,
+            trade_time,
             account_value,
             cash,
             return_rate,
@@ -56,14 +56,14 @@ class Report:
                 "None in [trade_date, account_value, cash, return_rate, turnover_rate, cost_rate, stock_value]"
             )
         # update report data
-        self.accounts[trade_date] = account_value
-        self.returns[trade_date] = return_rate
-        self.turnovers[trade_date] = turnover_rate
-        self.costs[trade_date] = cost_rate
-        self.values[trade_date] = stock_value
-        self.cashes[trade_date] = cash
+        self.accounts[trade_time] = account_value
+        self.returns[trade_time] = return_rate
+        self.turnovers[trade_time] = turnover_rate
+        self.costs[trade_time] = cost_rate
+        self.values[trade_time] = stock_value
+        self.cashes[trade_time] = cash
         # update latest_report_date
-        self.latest_report_date = trade_date
+        self.latest_report_time = trade_time
         # finish daily report update
 
     def generate_report_dataframe(self):
@@ -74,7 +74,7 @@ class Report:
         report["cost"] = pd.Series(self.costs)
         report["value"] = pd.Series(self.values)
         report["cash"] = pd.Series(self.cashes)
-        report.index.name = "date"
+        report.index.name = "trade_time"
         return report
 
     def save_report(self, path):
@@ -94,13 +94,13 @@ class Report:
 
         index = r.index
         self.init_vars()
-        for date in index:
+        for trade_time in index:
             self.update_report_record(
-                trade_date=date,
-                account_value=r.loc[date]["account"],
-                cash=r.loc[date]["cash"],
-                return_rate=r.loc[date]["return"],
-                turnover_rate=r.loc[date]["turnover"],
-                cost_rate=r.loc[date]["cost"],
-                stock_value=r.loc[date]["value"],
+                trade_time=trade_time,
+                account_value=r.loc[trade_time]["account"],
+                cash=r.loc[trade_time]["cash"],
+                return_rate=r.loc[trade_time]["return"],
+                turnover_rate=r.loc[trade_time]["turnover"],
+                cost_rate=r.loc[trade_time]["cost"],
+                stock_value=r.loc[trade_time]["value"],
             )
