+++ qlib/contrib/strategy/order_generator.py
@@ -4,8 +4,8 @@
 """
 This order generator is for strategies based on WeightStrategyBase
 """
-from ...backtest.position import Position
-from ...backtest.exchange import Exchange
+from ..backtest.position import Position
+from ..backtest.exchange import Exchange
 import pandas as pd
 import copy
