+++ qlib/data/data.py
@@ -117,6 +117,7 @@ class CalendarProvider(abc.ABC):
         flag = f"{freq}_sam_{freq_sam}_future_{future}"
         if flag in H["c"]:
             _calendar, _calendar_index = H["c"][flag]
+            return _calendar, _calendar_index
         else:
             flag_raw = f"{freq}_sam_{None}_future_{future}"
             if flag_raw in H["c"]:
@@ -125,6 +126,7 @@ class CalendarProvider(abc.ABC):
                 _calendar = np.array(self.load_calendar(freq, future))
                 _calendar_index = {x: i for i, x in enumerate(_calendar)}  # for fast search
                 H["c"][flag_raw] = _calendar, _calendar_index
+                
             if freq_sam is None:
                 return _calendar, _calendar_index
             else:
@@ -132,6 +134,7 @@ class CalendarProvider(abc.ABC):
                 _calendar_sam_index = {x: i for i, x in enumerate(_calendar_sam)}
                 H["c"][flag] = _calendar_sam, _calendar_sam_index
                 return _calendar_sam, _calendar_sam_index
+                
 
     def _uri(self, start_time, end_time, freq, future=False):
         """Get the uri of calendar generation task."""
@@ -541,8 +544,8 @@ class LocalCalendarProvider(CalendarProvider):
         with open(fname) as f:
             return [pd.Timestamp(x.strip()) for x in f]
 
-    def calendar(self, start_time=None, end_time=None, freq="day", future=False, freq_sam=None):
-        _calendar, _ = self._get_calendar(freq=freq, future=future)
+    def calendar(self, start_time=None, end_time=None, freq="day", freq_sam=None, future=False):
+        _calendar, _ = self._get_calendar(freq=freq, freq_sam=freq_sam, future=future)
         # strip
         if start_time:
             start_time = pd.Timestamp(start_time)
@@ -764,6 +767,7 @@ class ClientCalendarProvider(CalendarProvider):
         self.conn = conn
 
     def calendar(self, start_time=None, end_time=None, freq="day", future=False):
+        
         self.conn.send_request(
             request_type="calendar",
             request_content={
